DCWL-2003-v2 moved connectors
